requirements.txt
asgiref==3.8.1
Django==5.1.6
django-ordered-model==3.7.4
sqlparse==0.5.3
tzdata==2025.1
